fix(i18n): share locale week-info resolution with the date picker
Extract getLocaleWeekInfo/getFirstWeekday/getWeekendDays into
@stll/ui/lib/week (with the legacy Intl weekInfo accessor fallback). The
date picker and apps/web week.ts now resolve the week start and weekend
from one source, so the picker honors the legacy accessor instead of
silently falling back to Monday.
